requirements.txt
pyserial
flake8
